ISSD-1366 Refactor Routes
