--- workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/AccountDetailsPage/AccountHeaderButton.scss
@@ -1,34 +1,0 @@
-.account-details-header-right-content {
-	&-button {
-		&-container {
-			align-items: baseline;
-			display: flex;
-			gap: 0.25rem;
-			justify-content: center;
-		}
-
-		&-text {
-			font-size: 1rem;
-			font-weight: 400;
-			line-height: 1.5rem;
-
-			&-bold {
-				font-weight: 600;
-			}
-		}
-	}
-
-	&-container {
-		align-items: flex-start;
-		display: flex;
-		flex-direction: column;
-		gap: 1rem;
-	}
-
-	&-title {
-		color: #282934;
-		font-size: 1rem;
-		font-weight: 600;
-		line-height: 1.5rem;
-	}
-}
