+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PurchasedAppsDashboard/Apps/App.scss
@@ -1,0 +1,173 @@
+.app-details-page {
+	&-container {
+		margin-left: 80px;
+		width: 100%;
+	}
+
+	&-alert {
+		&-container {
+			background-color: #e6ebf5;
+			border: none;
+			border-radius: 0.5rem;
+			margin-bottom: 32px;
+		}
+
+		&-items-container {
+			align-items: center;
+			display: flex;
+			gap: 0.5rem;
+			justify-content: flex-start;
+		}
+
+		&-icon {
+			filter: invert(21%) sepia(25%) saturate(3697%) hue-rotate(199deg)
+				brightness(98%) contrast(88%);
+		}
+
+		&-text {
+			color: #1c3667;
+			font-size: 1rem;
+			font-weight: 400;
+			line-height: 1.5rem;
+		}
+	}
+
+	&-back-button {
+		background: none;
+		border: none;
+		border-radius: 6px;
+		color: #2b3a4b;
+		font-size: 0.875rem;
+		font-weight: 600;
+		height: 2rem;
+		line-height: 1rem;
+		margin-bottom: 1.5rem;
+		width: 7.625rem;
+
+		&-icon {
+			filter: invert(18%) sepia(3%) saturate(5893%) hue-rotate(172deg)
+				brightness(95%) contrast(85%);
+			height: 1rem;
+			width: 1rem;
+		}
+
+		&:hover {
+			background-color: #e6ebf5;
+		}
+
+		transition: background-color 0.3s;
+	}
+
+	&-app-info {
+		&-main-container {
+			align-items: center;
+			display: flex;
+			justify-content: space-between;
+			margin-bottom: 2.5rem;
+		}
+
+		&-logo {
+			border-radius: 0.5rem;
+			height: 3rem;
+			width: 3rem;
+		}
+
+		&-left-container {
+			align-items: center;
+			display: flex;
+			gap: 1rem;
+		}
+
+		&-title {
+			color: #282934;
+			font-size: 1.5rem;
+			font-weight: 600;
+			line-height: 1.75rem;
+		}
+
+		&-subtitle {
+			&-container {
+				align-items: center;
+				display: flex;
+				gap: 0.5rem;
+			}
+
+			&-icon {
+				height: 0.5rem;
+				width: 0.5rem;
+
+				&-published {
+					filter: invert(32%) sepia(12%) saturate(6208%)
+						hue-rotate(194deg) brightness(90%) contrast(87%);
+				}
+
+				&-hidden {
+					filter: invert(56%) sepia(4%) saturate(723%)
+						hue-rotate(202deg) brightness(91%) contrast(83%);
+				}
+
+				&-pending {
+					filter: invert(29%) sepia(88%) saturate(3491%)
+						hue-rotate(27deg) brightness(96%) contrast(101%);
+				}
+			}
+
+			&-text {
+				color: #54555f;
+				font-size: 0.875rem;
+				font-weight: 600;
+				line-height: 1.625rem;
+			}
+		}
+
+		&-buttons-container {
+			align-items: center;
+			display: flex;
+			gap: 1rem;
+		}
+
+		&-button {
+			&-preview-app-page {
+				background: none;
+				border: 1.5px solid #0b5fff;
+				border-radius: 0.375rem;
+				color: #0b5fff;
+				font-size: 0.875rem;
+				font-weight: 600;
+				height: 2rem;
+				line-height: 1rem;
+				width: 9rem;
+
+				&:hover {
+					background: #0b5fff;
+					color: #ffff;
+				}
+
+				transition: background 0.3s;
+				transition: color 0.3s;
+			}
+
+			&-manage {
+				background: #0b5fff;
+				border: none;
+				border-radius: 0.375rem;
+				color: #ffff;
+				font-size: 0.875rem;
+				font-weight: 600;
+				height: 2rem;
+				line-height: 1rem;
+				width: 5.688rem;
+
+				&-icon {
+					filter: invert(100%) sepia(0%) saturate(7463%)
+						hue-rotate(47deg) brightness(97%) contrast(99%);
+				}
+			}
+		}
+	}
+
+	&-navigation-bar {
+		margin-bottom: 32px;
+		max-width: max-content;
+	}
+}
