+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PurchasedSolutions/CreatedProjectCard.tsx
@@ -15,57 +15,54 @@ type CreatedProjectCardProps = {
 	product?: Product;
 };
 
-const CreatedProjectCard: React.FC<CreatedProjectCardProps> = ({product}) => {
-	return (
-		<div className="align-items-center d-flex flex-column h-100 justify-content-center purchased-solutions-container w-100">
-			<div className="border p-8 purchased-solutions-body rounded">
-				<div className="align-items-center d-flex flex-column justify-content-center">
-					<div className="mb-6">
-						<img
-							alt=""
-							className="gate-card-image"
-							src={createdProjectIcon}
-						/>
-					</div>
+const CreatedProjectCard: React.FC<CreatedProjectCardProps> = ({product}) => (
+	<div className="align-items-center d-flex flex-column h-100 justify-content-center purchased-solutions-container w-100">
+		<div className="border p-8 purchased-solutions-body rounded">
+			<div className="align-items-center d-flex flex-column justify-content-center">
+				<div className="mb-6">
+					<img
+						alt=""
+						className="gate-card-image"
+						src={createdProjectIcon}
+					/>
+				</div>
 
-					<div className="col-10 mb-2 mt-5 text-center">
-						<h1>
-							Your &nbsp;
-							<span className="created-project-cart-title">
-								{product?.name?.en_US}
-							</span>
-							&nbsp; project is being created now.
-						</h1>
-					</div>
+				<div className="col-10 mb-2 mt-5 text-center">
+					<h1>
+						Your &nbsp;
+						<span className="created-project-cart-title">
+							{product?.name?.en_US}
+						</span>
+						&nbsp; project is being created now.
+					</h1>
+				</div>
 
-					<div className="col-8 text-center">
-						<div>
-							<span>
-								Expect two emails in 10 minutes or less to
-								verify your project and extension environments
-								are ready.
-							</span>
-						</div>
+				<div className="col-8 text-center">
+					<div>
+						<span>
+							Expect two emails in 10 minutes or less to verify
+							your project and extension environments are ready.
+						</span>
 					</div>
+				</div>
 
-					<div className="mt-6 purchased-solutions-button-container">
-						<ClayButton
-							className="py-3"
-							onClick={() => {
-								window.location.href =
-									'https://www.liferay.com/pt/home';
-							}}
-						>
-							Return to Liferay.com
-							<span className="ml-3">
-								<ClayIcon symbol="order-arrow-right" />
-							</span>
-						</ClayButton>
-					</div>
+				<div className="mt-6 purchased-solutions-button-container">
+					<ClayButton
+						className="py-3"
+						onClick={() => {
+							window.location.href =
+								'https://www.liferay.com/pt/home';
+						}}
+					>
+						Return to Liferay.com
+						<span className="ml-3">
+							<ClayIcon symbol="order-arrow-right" />
+						</span>
+					</ClayButton>
 				</div>
 			</div>
 		</div>
-	);
-};
+	</div>
+);
 
 export default CreatedProjectCard;
