+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PurchasedSolutions/PurchasedSolutionsAccountSelection.tsx
@@ -3,7 +3,6 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import ClayAlert, {DisplayType} from '@clayui/alert';
 import ClayButton from '@clayui/button';
 import ClayForm, {ClayRadio} from '@clayui/form';
 import ClayLink from '@clayui/link';
@@ -17,7 +16,9 @@ import './PurchasedSolutions.scss';
 import ClaySticker from '@clayui/sticker';
 import classNames from 'classnames';
 
-import {getChannels, getOrderTypes, postOrder} from '../../utils/api';
+import {useMarketplaceContext} from '../../context/MarketplaceContext';
+import {Liferay} from '../../liferay/liferay';
+import {getOrderTypes, postOrder} from '../../utils/api';
 
 type Steps = {
 	page: 'accountCreation' | 'accountSelection' | 'projectCreated';
@@ -45,23 +46,7 @@ const PurchasedSolutionsAccountSelection: React.FC<PurchasedSolutionsccountSelec
 	const [radio, setRadio] = useState<RadioOption<Account>>();
 	const [orderType, setOrderType] = useState<OrderType>();
 	const [disabledButton, setDisabledButton] = useState<boolean>(false);
-	const [toastItems, setToastItems] = useState<
-		{message: string; title?: string; type: DisplayType}[]
-	>([]);
-
-	const renderToast = (message: string, title: string, type: DisplayType) => {
-		setToastItems([...toastItems, {message, title, type}]);
-	};
-
-	const [channel, setChannel] = useState<Channel>({
-		channelId: 0,
-		currencyCode: '',
-		externalReferenceCode: '',
-		id: 0,
-		name: '',
-		siteGroupId: 0,
-		type: '',
-	});
+	const {channel} = useMarketplaceContext();
 
 	const trialLenght =
 		orderInfo?.specifications &&
@@ -69,18 +54,6 @@ const PurchasedSolutionsAccountSelection: React.FC<PurchasedSolutionsccountSelec
 			(specification) =>
 				specification?.specificationKey === 'trial-length'
 		);
-
-	const renderToastMessage = () => {
-		renderToast(
-			'We are unable to start your trial. Please contact our sales team via email - sales@liferay.com',
-			'',
-			'danger'
-		);
-	};
-
-	const findChannelByName = (channels: Channel[], name: String) => {
-		return channels.find((channel: Channel) => channel.name === name);
-	};
 
 	const findOrderTypeByName = (
 		orderTypes: OrderType[],
@@ -93,20 +66,17 @@ const PurchasedSolutionsAccountSelection: React.FC<PurchasedSolutionsccountSelec
 	};
 
 	const fetchDataAndSetState = async () => {
-		const channels = await getChannels();
 		const orderTypes = await getOrderTypes();
 
-		if (!channels.length || !orderTypes.length) {
+		if (!channel || !orderTypes.length) {
 			setDisabledButton(true);
 
-			renderToastMessage();
-
-			return;
+			return Liferay.Util.openToast({
+				message:
+					'We are unable to start your trial. Please contact our sales team via email - sales@liferay.com',
+				type: 'danger',
+			});
 		}
-
-		const channel =
-			findChannelByName(channels, 'Marketplace Channel') || channels[0];
-		setChannel(channel);
 
 		const projectOrderType = findOrderTypeByName(
 			orderTypes,
@@ -141,7 +111,7 @@ const PurchasedSolutionsAccountSelection: React.FC<PurchasedSolutionsccountSelec
 	};
 
 	const onsubmit = async () => {
-		const payload: Order = {
+		await postOrder({
 			account: {
 				id: Number(radio?.value?.id),
 				type: radio?.value?.type as string,
@@ -168,152 +138,127 @@ const PurchasedSolutionsAccountSelection: React.FC<PurchasedSolutionsccountSelec
 			orderTypeId: Number(orderType?.id),
 			shippingAmount: 0,
 			shippingWithTaxAmount: 0,
-		};
-
-		await postOrder(payload);
+		});
 
 		setStep({page: 'projectCreated'});
 	};
 
 	return (
-		<>
-			<div className="align-items-center d-flex flex-column justify-content-center purchased-solutions-container">
-				<div className="border p-8 purchased-solutions-body rounded">
-					<span className="d-flex justify-content-center">
-						<Header description title="Account Selection" />
+		<div className="align-items-center d-flex flex-column justify-content-center purchased-solutions-container">
+			<div className="border p-8 purchased-solutions-body rounded">
+				<span className="d-flex justify-content-center">
+					<Header description title="Account Selection" />
+				</span>
+
+				<div className="mb-4">
+					<span>
+						{`Accounts available for `}
+
+						<strong>{currentUserAccount?.emailAddress}</strong>
+
+						{` (you)`}
 					</span>
+				</div>
 
-					<div className="mb-4">
-						<span>
-							{`Accounts available for `}
-
-							<strong>{currentUserAccount?.emailAddress}</strong>
-
-							{` (you)`}
-						</span>
-					</div>
-
-					<ClayForm>
-						<ClayForm.Group>
-							<div className="d-flex justify-content-between">
-								<div className="form-group mb-0 pr-3 w-100">
-									{accounts.map((account, index) => (
-										<div
-											className={classNames(
-												'align-items-center d-flex form-control justify-content-between mb-5 cursor-pointer',
-												{
-													fieldchecked:
-														radio?.index === index,
-												}
-											)}
-											key={index}
-											onClick={() =>
-												setRadio({
-													index,
-													value: account,
-												})
+				<ClayForm>
+					<ClayForm.Group>
+						<div className="d-flex justify-content-between">
+							<div className="form-group mb-0 pr-3 w-100">
+								{accounts.map((account, index) => (
+									<div
+										className={classNames(
+											'align-items-center d-flex form-control justify-content-between mb-5 cursor-pointer',
+											{
+												fieldchecked:
+													radio?.index === index,
 											}
-										>
-											<span className="align-items-center d-flex p-2">
-												<ClaySticker
-													shape="circle"
-													size="lg"
-												>
-													<ClaySticker.Image
-														alt="placeholder"
-														src={
-															account?.logoURL ??
-															emptyPictureIcon
-														}
-													/>
-												</ClaySticker>
-
-												<h5 className="mb-0 ml-3">
-													{account?.name}
-												</h5>
-											</span>
-
-											<div className="pr-4">
-												<ClayRadio
-													checked={
-														radio?.index === index
-													}
-													className="mr-5"
-													onChange={() =>
-														setRadio({
-															index,
-															value: account,
-														})
-													}
-													type="radio"
-													value="um"
-												/>
-											</div>
-										</div>
-									))}
-								</div>
-							</div>
-
-							<div>
-								<span>Not seeing a specific Account? </span>
-
-								<ClayLink href="http://help.liferay.com/">
-									Contact Support
-								</ClayLink>
-							</div>
-
-							<div className="mt-6 purchased-solutions-button-container">
-								<div className="align-items-center d-flex justify-content-between w-100">
-									<div>
-										<ClayButton
-											className="font-weight-bold"
-											displayType="unstyled"
-											onClick={() => {
-												setStep({
-													page: 'accountCreation',
-												});
-											}}
-										>
-											Cancel
-										</ClayButton>
-									</div>
-
-									<ClayButton
-										disabled={
-											!radio?.value ||
-											disabledButton ||
-											!orderInfo?.sku
-										}
+										)}
+										key={index}
 										onClick={() =>
-											orderInfo?.sku && onsubmit()
+											setRadio({
+												index,
+												value: account,
+											})
 										}
 									>
-										Continue
+										<span className="align-items-center d-flex p-2">
+											<ClaySticker
+												shape="circle"
+												size="lg"
+											>
+												<ClaySticker.Image
+													alt="placeholder"
+													src={
+														account?.logoURL ??
+														emptyPictureIcon
+													}
+												/>
+											</ClaySticker>
+
+											<h5 className="mb-0 ml-3">
+												{account?.name}
+											</h5>
+										</span>
+
+										<div className="pr-4">
+											<ClayRadio
+												checked={radio?.index === index}
+												className="mr-5"
+												onChange={() =>
+													setRadio({
+														index,
+														value: account,
+													})
+												}
+												type="radio"
+												value="um"
+											/>
+										</div>
+									</div>
+								))}
+							</div>
+						</div>
+
+						<div>
+							<span>Not seeing a specific Account? </span>
+
+							<ClayLink href="http://help.liferay.com/">
+								Contact Support
+							</ClayLink>
+						</div>
+
+						<div className="mt-6 purchased-solutions-button-container">
+							<div className="align-items-center d-flex justify-content-between w-100">
+								<div>
+									<ClayButton
+										className="font-weight-bold"
+										displayType="unstyled"
+										onClick={() => {
+											setStep({
+												page: 'accountCreation',
+											});
+										}}
+									>
+										Cancel
 									</ClayButton>
 								</div>
+
+								<ClayButton
+									disabled={
+										!radio?.value ||
+										disabledButton ||
+										!orderInfo?.sku
+									}
+									onClick={() => orderInfo?.sku && onsubmit()}
+								>
+									Continue
+								</ClayButton>
 							</div>
-						</ClayForm.Group>
-					</ClayForm>
-				</div>
+						</div>
+					</ClayForm.Group>
+				</ClayForm>
 			</div>
-			<ClayAlert.ToastContainer>
-				{toastItems?.map((alert, index) => (
-					<ClayAlert
-						autoClose={5000}
-						displayType={alert.type}
-						key={index}
-						onClose={() => {
-							setToastItems((prevItems) =>
-								prevItems.filter((item) => item !== alert)
-							);
-						}}
-						title={alert.title}
-					>
-						{alert.message}
-					</ClayAlert>
-				))}
-			</ClayAlert.ToastContainer>
-		</>
+		</div>
 	);
 };
